+++ src/Frontend/Dfe.Complete/Pages/Projects/TaskList/Tasks/ConfirmBankDetailsTask/ConfirmBankDetailsTask.cshtml.cs
@@ -1,4 +1,7 @@
+using Dfe.Complete.Application.Projects.Commands.TaskData;
+using Dfe.Complete.Constants;
 using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Domain.ValueObjects;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -8,10 +11,23 @@ namespace Dfe.Complete.Pages.Projects.TaskList.Tasks.ConfirmBankDetailsTask
     public class ConfirmBankDetailsTaskModel(ISender sender, IAuthorizationService authorizationService, ILogger<ConfirmBankDetailsTaskModel> logger)
     : BaseProjectTaskModel(sender, authorizationService, logger, NoteTaskIdentifier.ConfirmBankDetailsForGeneralAnnualGrantPaymentNeedToChange)
     {
+        [BindProperty]
+        public Guid? TasksDataId { get; set; }
+
+        [BindProperty]
+        public bool? BankDetailsChangingYesNo { get; set; }
         public override async Task<IActionResult> OnGetAsync()
         {
             await base.OnGetAsync();
+            TasksDataId = Project.TasksDataId?.Value;
+            BankDetailsChangingYesNo = TransferTaskData.BankDetailsChangingYesNo;
             return Page();
+        }
+        public async Task<IActionResult> OnPostAsync()
+        {
+            await Sender.Send(new UpdateConfirmBankDetailsTaskCommand(new TaskDataId(TasksDataId.GetValueOrDefault())!, BankDetailsChangingYesNo));
+            SetTaskSuccessNotification();
+            return Redirect(string.Format(RouteConstants.ProjectTaskList, ProjectId));
         }
     }
 }