+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/user-capabilities-regional-workers/regional-caseworker-services/regional-casework-services-team-leader-user.cy.ts
@@ -3,6 +3,7 @@ import {
     shouldBeAbleToViewMultipleMonthsOfProjects,
     shouldBeAbleToViewReportsLandingPage,
     shouldNotBeAbleToCreateAProject,
+    shouldNotBeAbleToSoftDeleteAProject,
     shouldNotHaveAccessToViewAndEditUsers,
     shouldNotHaveAccessToViewConversionURNsPage,
     shouldNotHaveAccessToViewHandedOverProjects,
@@ -92,7 +93,9 @@ describe("Capabilities and permissions of the regional casework services team le
         shouldBeAbleToViewMultipleMonthsOfProjects();
     });
 
-    it("Should be able to assign unassigned projects to users", () => {
+    // bug 245061
+    // + POST /Projects/InternalContacts/EditAssignedUser is intermittently not being called
+    it.skip("Should be able to assign unassigned projects to users", () => {
         navBar.goToYourTeamProjects();
         yourTeamProjects
             .filterProjects("Unassigned")
@@ -132,8 +135,8 @@ describe("Capabilities and permissions of the regional casework services team le
         shouldNotBeAbleToCreateAProject();
     });
 
-    it.skip("Should NOT be able to soft delete projects", () => {
-        // not implemented
+    it("Should NOT be able to soft delete projects", () => {
+        shouldNotBeAbleToSoftDeleteAProject(projectId)
     });
 
     it("Should NOT be able to view and edit local authorities", () => {