+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/user-capabilities-support/data-consumer-user.cy.ts
@@ -5,6 +5,7 @@ import {
     shouldNotBeAbleToAddAProjectNote,
     shouldNotBeAbleToAddAProjectTaskNote,
     shouldNotBeAbleToCreateAProject,
+    shouldNotBeAbleToSoftDeleteAProject,
     shouldNotHaveAccessToViewAndEditUsers,
     shouldNotHaveAccessToViewConversionURNsPage,
     shouldNotHaveAccessToViewHandedOverProjects,
@@ -92,8 +93,8 @@ describe("Capabilities and permissions of the data consumer user", () => {
         shouldNotBeAbleToAddAProjectTaskNote(projectId);
     });
 
-    it.skip("Should NOT be able to soft delete projects", () => {
-        // not implemented
+    it("Should NOT be able to soft delete projects", () => {
+        shouldNotBeAbleToSoftDeleteAProject(projectId);
     });
 
     it("Should NOT be able to view and edit local authorities", () => {